Change range to one-way instead of one-time
I think this would be better, because there are defintely use cases where one wants to change the range min, max etc. values.

However, i'm not familiar with your "goales", so it might be that you want to reject this and encourage people to use the range input directly without AMB, if they have such a use case.